fix(queryjson): change occurence type from short to int

+++ src/main/scala/fr/aphp/id/eds/requester/query/QueryBuilderBasicResource.scala
@@ -325,7 +325,7 @@ class QueryBuilderBasicResource(val qbConfigs: QueryBuilderConfigs = new QueryBu
     def getPatientListFilteredByOccurrenceNumber(dataframe: DataFrame,
                                                  groupByColumns: ListBuffer[String],
                                                  operator: String,
-                                                 n: Short) = {
+                                                 n: Int) = {
 
       if (logger.isInfoEnabled)
         logger.info(s"*** GroupByColumns: $groupByColumns  FILTER: 'count $operator $n'")
